Tests/LibWeb/Text/expected/HTML/close-event.txt

type: close
wasClean: true
code: 1000
reason: well hello friends 🐞
